% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_xy_function.R
\name{corr.two.data.sets.v2}
\alias{corr.two.data.sets.v2}
\title{Correlate two datasets}
\usage{
corr.two.data.sets.v2(
  x_dataset,
  y_dataset,
  common_column = "cell.line",
  y_factors,
  x_factors,
  y_factor_name = "y",
  x_factor_name = "x",
  protein_scale = FALSE
)
}
\arguments{
\item{protein_scale}{Logical, whether to scale protein before fitting.}
}
\description{
Function taken from protools2 and slightly altered, main change is the calculation of
 R2 value
}
