% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertubation_data_sim.R
\name{fit_skewed_dists}
\alias{fit_skewed_dists}
\title{Fit skewed distributions and estimate parameters}
\usage{
fit_skewed_dists(data_df, distributions = c("normal", "skew-normal", "skew-t"))
}
\arguments{
\item{data_df}{data.frame of data to be modelled. Each column represents a
separate distribution.}

\item{distributions}{String vector with distributions to fit.}
}
\value{
data.frame with estimated distribution parameters and AIC metric.
}
\description{
Fit skewed distributions and estimate parameters
}
