% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_construction.R
\name{make_coef_signatures}
\alias{make_coef_signatures}
\title{Markers from coefficients}
\usage{
make_coef_signatures(
  data_files_dir,
  corr_data_files = NULL,
  positive_coef,
  list_of_filtered_perts = NULL,
  scale = FALSE,
  compound_label = FALSE
)
}
\arguments{
\item{data_files_dir}{Directory containing regression analysis.}

\item{corr_data_files}{Character vector containing file names of correlation
data output.}

\item{positive_coef}{String indicating whether positive coefficients are
indicative of sensitivity or resistance. Options are "resistance" or "sensitivity"}

\item{list_of_filtered_perts}{List of vectors containing perturbagens which
passed QC thresholds.}

\item{scale}{Logical, whether to scale beta coefficients before signature
construction.}
}
\description{
Create signatures from coefficients (linear model derived).
}
