% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation_filter.R
\name{perturbation_filter}
\alias{perturbation_filter}
\title{Filter perturbations on summary statistics}
\usage{
perturbation_filter(
  summary_stats_df,
  gather = TRUE,
  min_val_cut,
  max_val_cut,
  sd_cut,
  iqr_cut,
  skew_cut,
  kurt_cut
)
}
\arguments{
\item{kurt_cut}{}
}
\description{
Filter perturbations on summary statistics
}
