% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation_filter.R
\name{perturbation_summary_stats}
\alias{perturbation_summary_stats}
\title{Compute perturbagen summary statistics}
\usage{
perturbation_summary_stats(pert.datasets, prot.datasets)
}
\arguments{
\item{prot.datasets}{}
}
\description{
Compute perturbagen summary statistics
}
