% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_final_signature_list.R
\name{signature_list_from_markers}
\alias{signature_list_from_markers}
\title{Title}
\usage{
signature_list_from_markers(
  markers_dir,
  all_stringencies = TRUE,
  marker_type = NULL,
  signature_id_suffix
)
}
\arguments{
\item{markers_dir}{Directory leading to marker sheet subfolders (for different
stringency levels).}

\item{all_stringencies}{Logical. Whether 3 marker stringency levels has been implemented.}

\item{marker_type}{String. Indicate whether markers are for crispr, rnai or drug response.}

\item{signature_id_suffix}{String. Id for signature, such as metric used in regression analysis.}
}
\value{
List containing identified markers of response (resistance/sensitivity) for
 each perturbation ~ proteomics dataset combo.
}
\description{
Title
}
