% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertubation_data_sim.R
\name{simulate_density}
\alias{simulate_density}
\title{Simulate skewed density}
\usage{
simulate_density(cp_params, sample_no = 500)
}
\arguments{
\item{cp_params}{String vector of centred parameters for sn package.}

\item{sample_no}{Integer. Number of samples to draw from simulated distribution.}
}
\value{
data.frame with sampled values, respective probability density and iteration id.
}
\description{
Simulate skewed density
}
