% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertubation_data_sim.R
\name{simulate_distribution}
\alias{simulate_distribution}
\title{Sample skewed distribution}
\usage{
simulate_distribution(cp_params, sample_no = 500)
}
\arguments{
\item{cp_params}{String vector of centred parameters for sn package.}

\item{sample_no}{Integer. Number of samples to draw from simulated distribution.}
}
\value{
List with 2 data.frames, simulated_dists contains sampled values.
 df_quants contains quantiles and median.
}
\description{
Sample skewed distribution
}
