% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_xy_function.R
\name{corr.two.data.sets.v2}
\alias{corr.two.data.sets.v2}
\title{Correlate two datasets}
\usage{
corr.two.data.sets.v2(
  x_dataset,
  y_dataset,
  common_column = "cell.line",
  y_factors,
  x_factors,
  y_factor_name = "y",
  x_factor_name = "x",
  threshold_test = "Corr",
  tumour_type_covar = FALSE
)
}
\arguments{
\item{x_dataset}{Intended to be independent x variable of linear function. Here,
the proteomics datasets.}

\item{y_dataset}{Intended to be the dependent y variable of linear function. Here,
the perturbation datasets.}

\item{common_column}{Cell line ids, for merging.}

\item{y_factors}{Perturbations.}

\item{x_factors}{Proteins}

\item{threshold_test}{String. Determines initial test to determine the dependence
between the two variables.}

\item{tumour_type_covar}{Logical. Whether to include tumour type (haem / solid)
as covariate in linear models.}

\item{protein_scale}{Logical. whether to scale protein before fitting.}
}
\description{
Function taken from protools2 and slightly altered, main change is the calculation of
 R2 value. Protein df scaling can be performed outside of funtion.
}
